rust: fix expecting errors where there are none
Due to some recent-ish changes, the expression parsing a list of changes now returns `Vec<String>` instead of `Result<Vec<String>, _>`.

This leads to issues where `collect::<Result<_, _>>()` is called instead of `collect()`, and a compile-time error. This fix simply adds `Vec<String>` to the list
of types that can be parsed without being wrapped in a result. `parse_without_error` probably should be recursive but heh.